pyproject.toml
[project]
name = "quickstart_etl"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster",
    "dagster-cloud",
    "dagster-gcp",
    "dagster-gcp-pandas",
    "httpx",
    "dagster-dbt",
    "dagster-mlflow",
    "dagster-docker",
    "dask[complete]",
    "dask-bigquery",
    "pyarrow",
    "boto3",
    "pandas",
    "matplotlib",
    "scikit-learn",
    "google-cloud-bigquery",
    "google-cloud-storage",
    "lightgbm",
    "xgboost",
    "tensorflow",
    "mlflow",
    "optuna",
    "great-expectations",
    "pandera",
    "feast[gcp]",
    "tsfresh",
    "holidays",
    "fastapi",
    "uvicorn",
    "gunicorn", # production ASGI server
    "evidently", # data drift and prediction drift detection
    "prometheus-client", # exposing metrics for monitoring
    "pre-commit", # git hooks for linting and formatting
    "ruff", # linter and formatter
    "dvc[gs]", # data version control with Google Cloud Storage support
    "yfinance",
]

# Gcloud CLI
# google kubernetes engine (gke)
# Vertex AI (vertex)
# grafana (docker or gcp marketplace)
# github-actions (for CI/CD)
# terraform (for infrastructure as code)
# we should run uv sync every time we add something here to keep dependencies up to date


[dependency-groups]
dev = [
    "dagster-dg-cli>=1.11.11",
    "dagster-webserver",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quickstart_etl"]

[tool.dagster]
module_name = "quickstart_etl.definitions"
code_location_name = "quickstart_etl"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "quickstart_etl"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["quickstart_etl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
